+++ src/player/Loader/ResourceLoader.ts
@@ -1,7 +1,7 @@
 import debug from "debug";
 import { IAudioContext } from "standardized-audio-context";
 import { Audioset, Clip } from "../../audioset";
-import { decodeAudioBuffer } from "../decodeAudioBuffer";
+import { decodeAudioBuffer } from "../Audio";
 
 const log = debug("atpls:resources");
