--- src/player/index.ts
@@ -1,1 +1,0 @@
-export { getActiveAudioContext } from "./AudioContext";
